backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pypdf==3.17.4
Flask-JWT-Extended==4.6.0
bcrypt==4.1.3
